requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask==2.0.1
requests==2.25.1
Flask-RESTful==0.3.9
python-dotenv==0.18.0
Flask-MySQLdb==0.2.0
APScheduler==3.7.0
requests==2.25.1
redis==3.5.3
mysql-connector==2.2.9